+++ tests/functional_tests/expected_values/bitly.py
@@ -1,4 +1,5 @@
 import fixtures
+import constants
 from authomatic.providers import oauth2
 
 
@@ -35,10 +36,53 @@ CONFIG = {
         conf.user_id,
         conf.user_username_reverse,
         conf.user_name,
+
+        # User info JSON keys
+        'status_code', 'data', 'apiKey', 'domain_options', 'member_since',
+        'enterprise_permissions', 'has_master', 'profile_image',
+        'share_accounts', 'numeric_id', 'account_login', 'account_type',
+        'account_id', 'primary', 'visible', 'is_delegated', 'auto_import_links',
+        'full_name', 'account_name', 'is_enterprise', 'tracking_domains',
+        'default_link_privacy', 'display_name', 'custom_short_domain', 'login',
+        'is_verified', 'profile_url', 'status_txt',
     ],
     # Case insensitive
     'content_should_not_contain': conf.no_phone + conf.no_birth_date +
                                   conf.no_email + conf.no_location +
                                   conf.no_gender + conf.no_locale +
                                   conf.no_first_name + conf.no_last_name,
+    # True means that any thruthy value is expected
+    'credentials': {
+        'token_type': None,
+        'provider_type_id': '2-2',
+        '_expiration_time': None,
+        'consumer_key': None,
+        'provider_id': None,
+        'consumer_secret': None,
+        'token': True,
+        'token_secret': None,
+        '_expire_in': True,
+        'provider_name': 'bitly',
+        'refresh_token': None,
+        'provider_type': 'authomatic.providers.oauth2.OAuth2',
+        'refresh_status': constants.CREDENTIALS_REFRESH_NOT_SUPPORTED,
+    },
+    # Testing changes after credentials refresh
+    # same: True
+    # not same: False
+    # don't test: None
+    'credentials_refresh_change': {
+        'token_type': True,
+        'provider_type_id': True,
+        '_expiration_time': True,
+        'consumer_key': True,
+        'provider_id': True,
+        'consumer_secret': True,
+        'token': True,
+        'token_secret': True,
+        '_expire_in': True,
+        'provider_name': True,
+        'refresh_token': True,
+        'provider_type': True,
+    },
 }