+++ tests/functional_tests/fixtures/__init__.py
@@ -8,6 +8,8 @@ from jinja2 import Environment, FileSystemLoader
 # Add path of the functional_tests_path package to PYTHONPATH.
 # Tis is necessary for the following imports to work when this module is
 # imported from the expected_values.* modules.
+import time
+
 FUNCTIONAL_TESTS_PATH = path.join(path.dirname(__file__), '..')
 sys.path.append(FUNCTIONAL_TESTS_PATH)
 
@@ -40,12 +42,17 @@ def render_login_result(result):
     """
 
     response = None
+    original_credentials = {}
+    refreshed_credentials = {}
     if result:
         response_message = ''
         if result.user:
             result.user.update()
             if result.user.credentials:
+                original_credentials.update(result.user.credentials.__dict__)
+                time.sleep(2)
                 response = result.user.credentials.refresh(force=True)
+                refreshed_credentials.update(result.user.credentials.__dict__)
 
         user_properties = ASSEMBLED_CONFIG.values()[0]['user'].keys()
         template = env.get_template('login.html')
@@ -53,7 +60,9 @@ def render_login_result(result):
                                providers=ASSEMBLED_CONFIG.keys(),
                                user_properties=user_properties,
                                error=result.error,
-                               credentials_response=response)
+                               credentials_response=response,
+                               original_credentials=original_credentials,
+                               refreshed_credentials=refreshed_credentials)
 
 
 def get_configuration(provider):